chore: use utils lib in nuxt-ssr app
(re)move duplicate code
add test lib to utils package
