+++ apps/meta-data-utils/src/tableQuery.ts
@@ -1,43 +1,32 @@
-import type {
-  IColumn,
-  ISchemaMetaData,
-  ITableMetaData,
-  KeyObject,
-} from "./types";
 import {
   isValueType,
+  isRefType,
   isArrayType,
   isFileType,
-  isRefType,
 } from "./fieldHelpers";
+import { ISchemaMetaData, ITableMetaData, IColumn, KeyObject } from "./types";
 
 const FILE_FRAGMENT = "{ id, size, extension, url }";
 
 export const buildRecordDetailsQueryFields = (
   schemas: Record<string, ISchemaMetaData>,
-  schemaName: string,
+  schemaId: string,
   tableId: string
 ): string => {
-  const schemaMetaData = schemas[schemaName];
+  const schemaMetaData = schemas[schemaId];
   const tableMetaData = schemaMetaData.tables.find(
-    (t: ITableMetaData) =>
-      t.id.toLocaleLowerCase() === tableId.toLocaleLowerCase()
+    (t: ITableMetaData) => t.id === tableId
   );
 
   const allColumns = tableMetaData?.columns;
   const dataColumns = allColumns
-    ?.filter((c) => !c.name.startsWith("mg_"))
+    ?.filter((c) => !c.id.startsWith("mg_"))
     .filter((c) => c.columnType !== "HEADING");
 
   const refTableQueryFields = (refColumn: IColumn): string => {
     const refTableMetaData = schemas[
-      refColumn.refSchema || schemaName
-    ].tables.find(
-      (t: ITableMetaData) =>
-        t.id.toLocaleLowerCase() ===
-        // @ts-ignore we know that refTable is not undefined
-        convertToPascalCase(refColumn.refTable).toLocaleLowerCase()
-    );
+      refColumn.refSchemaId || schemaId
+    ].tables.find((t: ITableMetaData) => t.id === refColumn.refTableId);
 
     const allRefColumns = refTableMetaData?.columns;
 
@@ -92,19 +81,19 @@ export const buildRecordDetailsQueryFields = (
 };
 
 export const buildRecordListQueryFields = (
-  tableName: string,
-  schemaName: string,
+  tableId: string,
+  schemaId: string,
   schemas: Record<string, ISchemaMetaData>
 ) => {
-  const keyFields = buildKeyFields(tableName, schemaName, schemas);
-  const tableMetaData = getTableMetaData(schemas[schemaName], tableName);
+  const keyFields = buildKeyFields(tableId, schemaId, schemas);
+  const tableMetaData = getTableMetaData(schemas[schemaId], tableId);
 
   if (tableMetaData === undefined) {
     throw new Error(
-      "buildRecordListQueryFields; tableMetaData is undefined for tableName " +
-        tableName +
+      "buildRecordListQueryFields; tableMetaData is undefined for tableId " +
+        tableId +
         " in schema " +
-        schemaName
+        schemaId
     );
   }
 
@@ -113,7 +102,6 @@ export const buildRecordListQueryFields = (
   // suggested list fields that are part of this tableType
   const additionalFields: any = [
     "id",
-    "name",
     "label",
     "description",
     "pid",
@@ -144,12 +132,12 @@ const fieldsToQueryString = (fields: string[][]): string => {
 };
 
 const buildKeyFields = (
-  tableName: string,
-  schemaName: string,
+  tableId: string,
+  schemaId: string,
   schemas: Record<string, ISchemaMetaData>
 ) => {
-  const schemaMetaData = schemas[schemaName];
-  const tableMetaData = getTableMetaData(schemaMetaData, tableName);
+  const schemaMetaData = schemas[schemaId];
+  const tableMetaData = getTableMetaData(schemaMetaData, tableId);
 
   const keyFields = tableMetaData.columns.reduce(
     (acc: any, column: IColumn) => {
@@ -157,20 +145,20 @@ const buildKeyFields = (
         if (isValueType(column)) {
           acc.push(column.id);
         } else if (isRefType(column)) {
-          if (!column.refTable) {
+          if (!column.refTableId) {
             throw new Error(
               "refTable is undefined for refColumn with id " +
                 column.id +
                 " in table " +
-                tableName +
+                tableId +
                 ""
             );
           } else {
             acc.push(column.id);
             acc.push(
               buildKeyFields(
-                column.refTable,
-                column.refSchema || schemaName,
+                column.refTableId,
+                column.refSchemaId || schemaId,
                 schemas
               )
             );
@@ -201,8 +189,8 @@ export const extractExternalSchemas = (schemaMetaData: ISchemaMetaData) => {
     ...new Set(
       schemaMetaData.tables.reduce((acc: string[], table: ITableMetaData) => {
         table.columns.forEach((column: IColumn) => {
-          if (column.refSchema) {
-            acc.push(column.refSchema);
+          if (column.refSchemaId) {
+            acc.push(column.refSchemaId);
           }
         });
         return acc;
@@ -213,31 +201,31 @@ export const extractExternalSchemas = (schemaMetaData: ISchemaMetaData) => {
 
 export const extractKeyFromRecord = (
   record: any,
-  tableName: string,
+  tableId: string,
   schemaId: string,
   schemas: Record<string, ISchemaMetaData>
 ) => {
   const schemaMetaData = schemas[schemaId];
-  const tableMetaData = getTableMetaData(schemaMetaData, tableName);
+  const tableMetaData = getTableMetaData(schemaMetaData, tableId);
 
   const key = tableMetaData.columns.reduce((acc: any, column: IColumn) => {
     if (column.key === 1 && record[column.id]) {
       if (isValueType(column)) {
         acc[column.id] = record[column.id];
       } else if (isRefType(column)) {
-        if (!column.refTable) {
+        if (!column.refTableId) {
           throw new Error(
             "refTable is undefined for refColumn with id " +
               column.id +
               " in table " +
-              tableName +
+              tableId +
               ""
           );
         } else {
           acc[column.id] = extractKeyFromRecord(
             record[column.id],
-            column.refTable,
-            column.refSchema || schemaId,
+            column.refTableId,
+            column.refSchemaId || schemaId,
             schemas
           );
         }
@@ -275,29 +263,16 @@ export const buildFilterFromKeysObject = (keys: KeyObject) => {
 
 export const getTableMetaData = (
   schemaMetaData: ISchemaMetaData,
-  tableName: string
+  tableId: string
 ): ITableMetaData => {
   const tableMetaData = schemaMetaData.tables.find(
-    (t: ITableMetaData) =>
-      t.name.toLocaleLowerCase() === tableName.toLocaleLowerCase()
+    (t: ITableMetaData) => t.id === tableId
   );
 
   if (tableMetaData === undefined) {
-    const msg = "ERROR: tableMetaData is undefined for tableName " + tableName;
+    const msg = "ERROR: tableMetaData is undefined for tableId " + tableId;
     console.log(msg);
     throw new Error(msg);
   }
   return tableMetaData;
 };
-
-function convertToPascalCase(string: string): string {
-  const words = string.trim().split(/\s+/);
-  let result = "";
-  words.forEach((word: string) => {
-    result += word.charAt(0).toUpperCase();
-    if (word.length > 1) {
-      result += word.slice(1);
-    }
-  });
-  return result;
-}
