+++ apps/meta-data-utils/test/tableQuery.test.ts
@@ -1,11 +1,8 @@
 import { describe, it, expect } from "vitest";
+import { catalogueMetadata, catalogueOntologies, petStoreMetadata } from "./test-resources/metadata";
+import { buildRecordDetailsQueryFields } from "../src/tableQuery";
 
-import { buildRecordDetailsQueryFields } from "./tableQuery";
-import {
-  petStoreMetadata,
-  catalogueMetadata,
-  catalogueOntologies,
-} from "./test-resources/metadata";
+
 
 describe("buildRecordDetailsQueryFields", () => {
   it("should return the query fields simple schema (pet store)", () => {
