+++ apps/nuxt3-ssr/pages/[schema]/ssr-catalogue/variables/[variable]/index.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import variableQuery from "~~/gql/variable";
 import type { IVariable, IVariableMappings } from "~/interfaces/types";
+import { buildFilterFromKeysObject } from "meta-data-utils";
 const config = useRuntimeConfig();
 const route = useRoute();
