+++ apps/nuxt3-ssr/utils/headingUtils.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
 
 import { toHeadings, toSectionsMap } from "./headingUtils";
-import { ITableMetaData } from "~/interfaces/types";
+import { ITableMetaData } from "meta-data-utils";
 
 describe("toHeadings", () => {
   it("should return only the heading columns", () => {
